refactor

+++ include/franklin/spinlock.h
@@ -1,13 +1,14 @@
 #ifndef _SPIN_
 #define _SPIN_ 1
 
+#include <stdint.h>
 
-typedef unsigned int lock;
-unsigned int spinlock;
+typedef uint32_t lock;
+uint32_t spinlock;
 
 void acquire(lock*);
 void release(lock*);
 void init_lock(lock*);
-int trylock(lock*);
+uint8_t trylock(lock*);
 
 #endif
